@@ -1,68 +1,77 @@
+import 'package:booking_hotel/pages/onboardscreen.dart';
 import 'package:flutter/material.dart';
-import 'package:booking_hotel/pages/home_page.dart';
 
-class WelcomePage extends StatefulWidget {
-  const WelcomePage({Key? key}) : super(key: key);
+class WelcomeScreen extends StatelessWidget {
+  const WelcomeScreen({super.key});
 
-  @override
-  State<WelcomePage> createState() => _WelcomePageState();
-}
-
-class _WelcomePageState extends State<WelcomePage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: Center(
+    return Container(
+      decoration: const BoxDecoration(
+        image: DecorationImage(
+            image: AssetImage(
+                "assets/onboard/welcome.jpg"), // Đảm bảo đường dẫn đúng
+            fit: BoxFit.cover,
+            opacity: 0.7),
+      ),
+      child: Material(
+        color: Colors.transparent,
+        child: SafeArea(
           child: Padding(
-            padding: const EdgeInsets.all(16),
+            padding: const EdgeInsets.symmetric(vertical: 65, horizontal: 25),
             child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
               children: [
-                const Spacer(),
-                Image.asset(
-                  'assets/illustration.png',
-                ),
-                const SizedBox(height: 40),
                 const Text(
-                  "Explore your journey \nonly with us",
-                  textAlign: TextAlign.center,
+                  "Enjoy",
                   style: TextStyle(
-                    fontSize: 26,
+                    color: Colors.white,
+                    fontSize: 35,
                     fontWeight: FontWeight.bold,
+                    letterSpacing: 1.5,
                   ),
                 ),
-                const SizedBox(height: 20),
-                const Text(
-                  "All your vacations destinations are here,\nenjoy your holiday",
-                  textAlign: TextAlign.center,
+                const SizedBox(
+                  height: 2,
+                ),
+                Text(
+                  "the life!",
                   style: TextStyle(
-                    color: Colors.black54,
-                    fontSize: 16,
+                    color: Colors.white.withOpacity(0.9),
+                    fontSize: 35,
+                    fontWeight: FontWeight.w400,
+                    letterSpacing: 1.5,
                   ),
                 ),
-                const Spacer(),
-                Padding(
-                  padding: const EdgeInsets.only(bottom: 30),
-                  child: SizedBox(
-                    width: double.maxFinite,
-                    child: ElevatedButton(
-                      onPressed: () {
-                        Navigator.pushReplacement(
-                          context,
-                          MaterialPageRoute(
-                            builder: (context) => const HomePage(),
-                          ),
-                        );
-                      },
-                      style: ElevatedButton.styleFrom(
-                        elevation: 0,
-                        shape: const StadiumBorder(),
-                        padding: const EdgeInsets.symmetric(
-                          vertical: 15,
-                          horizontal: 8.0,
-                        ),
-                      ),
-                      child: const Text("Get Started"),
+                const SizedBox(
+                  height: 12,
+                ),
+                Text(
+                  "Khám phá và đặt phòng khách sạn một cách dễ dàng và nhanh chóng với ứng dụng của chúng tôi. Dễ sử dụng và tiện lợi, giúp bạn tìm kiếm các khách sạn phổ biến và đặt chỗ một cách linh hoạt. Bắt đầu cuộc hành trình của bạn ngay hôm nay!",
+                  style: TextStyle(
+                      color: Colors.white.withOpacity(0.7),
+                      fontSize: 16,
+                      letterSpacing: 1.2),
+                ),
+                const SizedBox(
+                  height: 30,
+                ),
+                InkWell(
+                  onTap: () {
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => const OnboardScreen()));
+                  },
+                  child: Ink(
+                    padding: const EdgeInsets.all(15),
+                    decoration: BoxDecoration(
+                        color: Colors.white,
+                        borderRadius: BorderRadius.circular(20)),
+                    child: const Icon(
+                      Icons.arrow_forward_ios,
+                      color: Colors.black54,
+                      size: 20,
                     ),
                   ),
                 )
