@@ -1,26 +1,29 @@
-import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
+import 'package:booking_hotel/pages/order_screen.dart';
 import 'package:booking_hotel/pages/welcome_page.dart.dart';
-
-void main() {
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:persistent_shopping_cart/persistent_shopping_cart.dart';
+void main() async{
+  await PersistentShoppingCart().init();
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
-
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return GetMaterialApp(
       title: 'Travel App',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+       backgroundColor: Colors.white,
         textTheme: GoogleFonts.mulishTextTheme(
           Theme.of(context).textTheme,
         ),
       ),
-      home: const WelcomePage(),
+      home: WelcomeScreen(),
+    
     );
   }
 }
