@@ -1,39 +1,78 @@
+import 'package:booking_hotel/consts.dart';
+import 'package:booking_hotel/pages/cart_page.dart';
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:get_storage/get_storage.dart';
 import 'package:ionicons/ionicons.dart';
 import 'package:booking_hotel/widgets/custom_icon_button.dart';
 import 'package:booking_hotel/widgets/location_card.dart';
 import 'package:booking_hotel/widgets/nearby_places.dart';
 import 'package:booking_hotel/widgets/recommended_places.dart';
-import 'package:booking_hotel/widgets/tourist_places.dart';
 
-class HomePage extends StatelessWidget {
+class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
 
+  @override
+  State<HomePage> createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  String username = 'Unknown';
+
+  @override
+  void initState() {
+    super.initState();
+    loadUserData();
+  }
+
+  Future<void> loadUserData() async {
+    final box = GetStorage();
+    final userData = box.read('user');
+    setState(() {
+      username = userData != null ? userData['username'] ?? 'Unknown' : 'Unknown';
+    });
+  }
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         elevation: 0,
         backgroundColor: Colors.transparent,
-        foregroundColor: Colors.black,
+        foregroundColor: const Color.fromARGB(255, 57, 43, 43),
         title: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             const Text("Good Morning"),
             Text(
-              "Tetteh Jeron Asiedu",
+              "$username",
               style: Theme.of(context).textTheme.labelMedium,
             ),
           ],
         ),
-        actions: const [
-          CustomIconButton(
-            icon: Icon(Ionicons.search_outline),
+        actions: [
+          IconButton(
+            onPressed: () {
+              Navigator.push(context,
+                  MaterialPageRoute(builder: (context) => CartScreen()));
+            },
+            color: kWhiteColor,
+            icon: Icon(
+              FontAwesomeIcons.search,
+              color: kTextColor,
+            ),
           ),
           Padding(
             padding: EdgeInsets.only(left: 8.0, right: 12),
-            child: CustomIconButton(
-              icon: Icon(Ionicons.notifications_outline),
+            child: IconButton(
+              onPressed: () {
+                Navigator.push(context,
+                    MaterialPageRoute(builder: (context) => CartScreen()));
+              },
+              color: kWhiteColor,
+              icon: Icon(
+                FontAwesomeIcons.cartShopping,
+                color: kTextColor,
+              ),
             ),
           ),
         ],
@@ -47,7 +86,7 @@ class HomePage extends StatelessWidget {
           const SizedBox(
             height: 15,
           ),
-          const TouristPlaces(),
+          // const TouristPlaces(),
           // CATEGORIES
           const SizedBox(height: 10),
           Row(
@@ -61,7 +100,7 @@ class HomePage extends StatelessWidget {
             ],
           ),
           const SizedBox(height: 10),
-          const RecommendedPlaces(),
+          RecommendedPlaces(),
           const SizedBox(height: 10),
           Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -75,29 +114,6 @@ class HomePage extends StatelessWidget {
           ),
           const SizedBox(height: 10),
           const NearbyPlaces(),
-        ],
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        type: BottomNavigationBarType.fixed,
-        showSelectedLabels: false,
-        showUnselectedLabels: false,
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Ionicons.home_outline),
-            label: "Home",
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Ionicons.bookmark_outline),
-            label: "Bookmark",
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Ionicons.ticket_outline),
-            label: "Ticket",
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Ionicons.person_outline),
-            label: "Profile",
-          )
         ],
       ),
     );
